outsource command functions
