+++ .prettierrc
@@ -1,1 +1,3 @@
-{}
+{
+  "singleQuote": true
+}
